[display][api-types-cpp] Add IsValid() and DebugAssertIsValid() to DisplayTiming.
DisplayTiming is a bag of display timing parameters, but with some
rep invariants for each parameter (and derived timing parameters,
such as refresh rate and horizontal / vertical totals).

We define DisplayTiming as a struct (as opposed to a class), so
that we can benefit in performance (when initialized) and readability
(since we can use aggregate initialization). But for structs there is
no enforced rep invariant.

This change adds IsValid() and DebugAssertIsValid() to DisplayTiming's
public API, so that clients may choose to check the validity of the
timing parameters.

Test: Added DisplayTimingIsValid tests to display-timing-test
Change-Id: Ia29647e6a411f9028ce5e7d224fdc7a8e504bc8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952177
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Reviewed-by: Victor Costan <costan@google.com>