[botanist] When retrying flashing, read from all serial sockets
We've observed these retries get attempted, but they appear to
subsequently fail.  In particular, it seems that something is causing
the serial output from the mux to stop, so botanist never sees the
serial logline indicating that the device has successfully booted.

I hypothesize that botanist failing to read from the socket it writes to
is maybe causing the serial mux implemented by dms to block when it
would otherwise attempt to write serial output from the device to other
clients with open sockets.  Between the previous bug which would cause
us to fail to close the socket entirely and a bug where failing to read
from the socket causes us to block serial output to all mux clients, we
were never seeing serial logs after partway through reboot after the
hard power-cycle.

Let's start by trying to be a better-behaved mux client, and separately
we can look into if it makes sense to make the mux drop lines to clients
that aren't keeping up, or if we'd rather keep things blocking and
demand well-behaved serial clients.

Bug: b/289078335
Change-Id: Icb15bc0a518dbe94c480f182f055a1cfb71f87fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952292
Commit-Queue: Drew Fisher <zarvox@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>