[ffx] Fix a few detached tasks not returning ()
The target version of Task::detach doesn't allow you to use a future
that returns anything other than (). We're about to add this requirement
to the host side as well for consistency so we need to fix the code that
uses that "feature".

Change-Id: Ib4afc121175f9f8ba77a17c19304ffe07c7fcffc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952189
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>