[kernel][tracing] Fix an issue with mutex spin-tracing.
Fix an issue with kernel spin-tracing used to instrument the spin
phase of adaptive mutexes which could cause the "blocked-after" flag
to be incorrectly reported.

After a thread gives up on spinning on an adaptive mutex, if
spin-tracing is enabled in the build, the thread logs a trace record
indicating how much time was spent spinning on the mutex before it
gave up and decided to block.  There is a flag associated with this
recorded called "blocked-after" which indicates whether or not the
thread needed to become blocked after the spin phase in order to
obtain the mutex.

In the past, if the thread managed to acquire the mutex during the
spin phase, the flag would be clear (no block was needed) and set
otherwise.  The trouble with this approach is that after the thread
gives up on spinning, it needs to obtain the global thread lock before
it actually blocks, and once inside the lock, it makes one final
attempt to obtain the mutex before blocking.  If it succeeds at this
point, it does not actually block, even though the spin-contention
trace event it just produced says that it will block.

Change the code to record the time at which the spinning behavior
ended, but don't actually mark whether the thread "blocked after"
until after we have entered the GTL and determined if we are actually
going to block.

Change-Id: I75beee2f54cf2dacb15c918d6fefd9b82a85cab9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952691
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>