[sandbox] Add Routable trait
The Routable trait moves the routing logic into a trait impl
instead of keeping it in an anyonmous function. This will let
us call route without having to clone the underlying object.

Also, replace some Routers with dictionaries in the sandbox.
A dictionary can always be turned into a Router, so this should
happen as late as possible.

Change-Id: Ibdd4e15b21130a83e5cd84eecf6dc9b81fd7945a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950094
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>