[audio][bluetooth] Return driver_transfer_bytes
The fuchsia audio driver definition requires the property
`driver_transfer_bytes` to be returned as part of the
RingBufferProperties table, and for this value to be non-
zero.

This CL updates the fuchsia-audio-device Bluetooth audio
library to be compliant, by reserving the size of a packet.
Note that this may not be an accurate measure of the actual
`driver_transfer_bytes`, but it seems better than the
alternative, which was to (for now) simply return 1 byte.

Furthermore, fuchsia-audio-device now reserves this
additional space, when allocating a ring-buffer.

With this change, a test case in bt-hfp-audio-gateway-tests
had to be reworked, because it relied upon an exact ring-
buffer size, compared to the number of MSBC frames and
notifications per ring-buffer.

After this CL, the audio_driver_tests suite can be updated
to require correct ring-buffer size-reservation behavior of
all drivers (subsequent CL).

Bug: fxbug.dev/123475

Test: updated unit tests; fx test fuchsia-audio-device-tests
Test: fx test bt-hfp-audio-gateway-tests

Change-Id: I0c72f3146e8d4730d863208c03d2c8706b7b694b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947544
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>