[audio] Add logging for SetVolume
For SetVolume calls, log even when this represents no
change. For both SetVolume and SetMute, if this represents
a change, also log the previous value and the number of
clients that will be notified.

This does not increase the number of lines logged on
shipping product builds.

Change-Id: I9398a22a9d74f6cebf68719e69a5cf21f0bda362
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/835217
Reviewed-by: Martin Puryear <mpuryear@google.com>
Fuchsia-Auto-Submit: Martin Puryear <mpuryear@google.com>
Reviewed-by: Cam Lloyd <camlloyd@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>