[netcfg] Remove AlreadyExists from name generation
An earlier phase of the project assumed that once an interface had a
name generated by netcfg, that any other interface that tried to use
that name would be rejected, unless the new interface had the same topo
path or MAC address.

This has now been adjusted and this code that checks for this assumption
can be removed. Devices should attempt to be added even if another
interface might have the same name. Netstack can reject at the point of
interface installation since it knows the current state of the
interfaces active.

Bug: 135087
Bug: 56559
Change-Id: I6f5669075993a2d4372412d328f04e68bfdb203a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952665
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>