[netcfg] Add name to AlreadyExists devices error
Include the interface name in the AlreadyExists device error when a
device fails to be added to an interface because an interface with the
name already exists.

We will be able to use this field to find the PersistentIdentifier for
the interface that already exists in the Netstack, then determine
whether we should remove an interface to install the new one or simply
reject the new interface.

Bug: 56559
Bug: 135087
Bug: 309010918
Change-Id: I0a12a0529a23cb6cb80a3513c45a0b23b56e0aab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952670
Reviewed-by: Zoey Bush <zmbush@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>