[bazel-sdk] better cleanup of build configs
When work-on exited it would try to restore the original build config
file. However, if there was a previous crash the file would not have
been deleted which was causing confusion.

This CL removes the original build configuration unless a user specifies
to keep it.

Change-Id: If4ed36321727fc18761c0a28154150c63a0afc64
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/954054
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>