[netcfg] Add PersistentId to InterfaceState
PersistentId will be needed in removing the temporary interface naming
policy. It will be used to determine whether the new interfaces are the
same logical interfaces as existing interfaces that may be active on the
network.

Additionally, return PersistentId from the generate_stable_name fn so
that it can be saved in the InterfaceState. Currently, it is stored as
an option and we can drop the option once we are only using
`generate_stable_name` and not `generate_temporary_name`.

Bug: b/309010918
Bug: 56559
Bug: 135087
Change-Id: I3cbd567c9d1495f1f0a0c73512a5df75d2fa03f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952815
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>