[power-manager] Delay reconnect attempts for ActivityProvider
PowerManager has logic to handle disconnects from the
fuchsia.ui.activity Listener stream by attempting to re-establish a new
channel through the Provider interface. In cases where the disconnect is
caused by a device reboot and both the Listener and Provider interfaces
are closed, it's possible the Provider interface closes after the
Listener interface but not before a request is made by activity_handler
to establish a new Listener stream. This CL introduces a short delay
before attempting to reestablish a connection to
fuchsia.ui.activity/Provider to mitigate the chances of that happening.

Bug: b/311783026
Change-Id: Iccd7f15013ab41abe22522a071acf3b0fffa0ab5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953532
Commit-Queue: Alex Wong <alexdwong@google.com>
Reviewed-by: Onath Dillinger <claridge@google.com>