[scsi][ufs] Support SCSI UNMAP command
This CL uses the inquiry command to determine if the device supports
the unmap command and, if supported, performs the unmap command.

For now, I am only going to apply the trim(unmap) command handling to
the UFS driver and leave the ums and virtio to a separate patch.

Test: ufs-unit-tests disk-test
Change-Id: Ibda343a69253c63691001a8ddd30dbd6e4907681
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930713
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Christopher Anderson <cja@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>