[starnix] Add the syslog(2) syscall
This change doesn't implement the functionality we need from this
syscall just yet. It just provides the basic handling of the syscall,
which turns out is sufficient to pass the gvisor tests, as well as
the basic checks for capabilities in the actions that require privilege
as documented on the Linux manual.

Bug: 314198813
Change-Id: I65df3540bdbff9ba8ecebb4513a2a5b6f9889316
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/954216
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>