[ffx][usb_fastboot_discovery]: only emit for Devices in fastboot
A usb interface can be discovered to be a candidate to be in fastboot
mode but without actually being in fastboot mode _yet_. This is a race
condition where the device is available on the bus but is the Target is
not yet initialized and ready to accept Fastboot commands.

Now introduce a checker trait to determine if the interface with the
provided serial is actually running in fastboot mode or not. The trait
is used for testing purposes but could be extended if need be.

Change-Id: I32260eb5172bc24999dce6dd15f423b89fd114b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953448
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>