[components] ELF runner can track all currently running components
In order to report the memory usage of each ELF component, the ELF
runner needs to be able to keep track of the ELF components that it is
currently running. Previously there's no way to do that short of relying
on fragile mapping from component to jobs. This CL adds a ComponentSet
type to weakly hold the set of running components.

Instead of the runner::component::Controller just taking ownership of
the ElfComponent, it is now ref-counted, with ComponentSet holding onto
a weak pointer.

Bug: b/307580082
Change-Id: I69da7409b1fa1a4adbc66be4c853fe4597160d8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/955713
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>