[tools/build] Use Build API script to get modules.
Introduce a new BuildApiClient struct that can be used to
retrieve the content of build API module files by invoking
the script at //build/api/client instead of reading the
GN-generated files directly.

This indirection will allow the introduction of
Bazel-generated outputs in the result in the future with
minimal changes to tools/build.

+ build/api/client: Fix a bug that prevented the script
  from working on MacOS (where OSTYPE is darwin22).

Bug: 134999
Change-Id: Ib71fa5be2d030b3f209c03a0cb6aa8b0b01ddd7f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/932212
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>