[netcfg] Remove persistent naming config
The naming policy specified by configuration enables the creation of
stable and consistent interface naming. The file-backed structure was
created with the intention to provide consistent naming, but this is no
longer necessary with the configuration-based naming policies.

The file-backed structure and the policy of looking up the name for
interfaces that were already known made it difficult to enact new
naming policy without having this file reset (e.g. fresh fx flash).

Bug: 135087
Fixed: 137183
Change-Id: Idacd11e2b9e4b2cbec1df91a75fbafb01eba58ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956168
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>