[audio][drivers] codec.fidl param name (2 of 2)
As part 2 of a two-part transition, remove the deprecated
parameter `start_time` that was previously returned from
fuchsia.hardware.audio.Codec/Stop. Because all drivers
are still in-tree, and all Codec drivers are serviced by
a single helper library(!), this can actually be done in
one step. Because it is simply a renaming of a parameter
(no change to its type), this is actually already ABI-
compatible.

Test: fx test -ouv audio_driver_tests

Change-Id: Ic227846d0142802c55e776ec29e316edd479a9bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953702
API-Review: Dale Sather <dalesat@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>